refactor: RestaurantWithDistanceDao 제거하고 SimpleResponse로 통일 (#548)
